+++ source/extensions/isaacsim.robot.policy.examples/isaacsim/robot/policy/examples/robots/spot.py
@@ -7,24 +7,18 @@
 # license agreement from NVIDIA CORPORATION is strictly prohibited.
 #
 
-import io
-from typing import List, Optional
+from typing import Optional
 
-import carb
 import numpy as np
 import omni
 import omni.kit.commands
-import torch
-from isaacsim.core.prims import SingleArticulation
-from isaacsim.core.utils.prims import define_prim, get_prim_at_path
-from isaacsim.core.utils.rotations import euler_to_rot_matrix, quat_to_euler_angles, quat_to_rot_matrix
-from isaacsim.core.utils.stage import get_current_stage
+from isaacsim.core.utils.rotations import quat_to_rot_matrix
 from isaacsim.core.utils.types import ArticulationAction
+from isaacsim.robot.policy.examples.controllers import PolicyController
 from isaacsim.storage.native import get_assets_root_path
-from pxr import Gf
 
 
-class SpotFlatTerrainPolicy:
+class SpotFlatTerrainPolicy(PolicyController):
     """The Spot quadruped"""
 
     def __init__(
@@ -46,43 +40,19 @@ class SpotFlatTerrainPolicy:
             orientation {np.ndarray} -- orientation of the robot
 
         """
-        self._stage = get_current_stage()
-        self._prim_path = prim_path
-        prim = get_prim_at_path(self._prim_path)
-
         assets_root_path = get_assets_root_path()
-        if not prim.IsValid():
-            prim = define_prim(self._prim_path, "Xform")
-            if usd_path:
-                prim.GetReferences().AddReference(usd_path)
-            else:
-                if assets_root_path is None:
-                    carb.log_error("Could not find Isaac Sim assets folder")
+        if usd_path == None:
+            usd_path = assets_root_path + "/Isaac/Robots/BostonDynamics/spot/spot.usd"
 
-                asset_path = assets_root_path + "/Isaac/Robots/BostonDynamics/spot/spot.usd"
+        super().__init__(name, prim_path, usd_path, position, orientation)
 
-                prim.GetReferences().AddReference(asset_path)
-
-        self.robot = SingleArticulation(
-            prim_path=self._prim_path, name=name, position=position, orientation=orientation
+        self.load_policy(
+            assets_root_path + "/Isaac/Samples/Policies/Spot_Policies/spot_policy.pt",
+            assets_root_path + "/Isaac/Samples/Policies/Spot_Policies/spot_env.yaml",
         )
-
-        self._dof_control_modes: List[int] = list()
-
-        # Policy
-        file_content = omni.client.read_file(
-            assets_root_path + "/Isaac/Samples/Quadruped/Spot_Policies/spot_policy.pt"
-        )[2]
-        file = io.BytesIO(memoryview(file_content).tobytes())
-
-        self._policy = torch.jit.load(file)
-        self._base_vel_lin_scale = 1
-        self._base_vel_ang_scale = 1
         self._action_scale = 0.2
-        self._default_joint_pos = np.array([0.1, -0.1, 0.1, -0.1, 0.9, 0.9, 1.1, 1.1, -1.5, -1.5, -1.5, -1.5])
         self._previous_action = np.zeros(12)
         self._policy_counter = 0
-        self._decimation = 10
 
     def _compute_observation(self, command):
         """
@@ -107,26 +77,24 @@ class SpotFlatTerrainPolicy:
 
         obs = np.zeros(48)
         # Base lin vel
-        obs[:3] = self._base_vel_lin_scale * lin_vel_b
+        obs[:3] = lin_vel_b
         # Base ang vel
-        obs[3:6] = self._base_vel_ang_scale * ang_vel_b
+        obs[3:6] = ang_vel_b
         # Gravity
         obs[6:9] = gravity_b
         # Command
-        obs[9] = self._base_vel_lin_scale * command[0]
-        obs[10] = self._base_vel_lin_scale * command[1]
-        obs[11] = self._base_vel_ang_scale * command[2]
+        obs[9:12] = command
         # Joint states
         current_joint_pos = self.robot.get_joint_positions()
         current_joint_vel = self.robot.get_joint_velocities()
-        obs[12:24] = current_joint_pos - self._default_joint_pos
+        obs[12:24] = current_joint_pos - self.default_pos
         obs[24:36] = current_joint_vel
         # Previous Action
         obs[36:48] = self._previous_action
 
         return obs
 
-    def advance(self, dt, command):
+    def forward(self, dt, command):
         """
         Compute the desired torques and apply them to the articulation
 
@@ -137,27 +105,10 @@ class SpotFlatTerrainPolicy:
         """
         if self._policy_counter % self._decimation == 0:
             obs = self._compute_observation(command)
-            with torch.no_grad():
-                obs = torch.from_numpy(obs).view(1, -1).float()
-                self.action = self._policy(obs).detach().view(-1).numpy()
+            self.action = self._compute_action(obs)
             self._previous_action = self.action.copy()
 
-        action = ArticulationAction(joint_positions=self._default_joint_pos + (self.action * self._action_scale))
+        action = ArticulationAction(joint_positions=self.default_pos + (self.action * self._action_scale))
         self.robot.apply_action(action)
 
         self._policy_counter += 1
-
-    def initialize(self, physics_sim_view=None) -> None:
-        """
-        Initialize robot the articulation interface, set up drive mode
-        """
-        self.robot.initialize(physics_sim_view=physics_sim_view)
-        self.robot.get_articulation_controller().set_effort_modes("force")
-        self.robot.get_articulation_controller().switch_control_mode("position")
-        self.robot._articulation_view.set_gains(np.zeros(12) + 60, np.zeros(12) + 1.5)
-
-    def post_reset(self) -> None:
-        """
-        Post reset articulation
-        """
-        self.robot.post_reset()